+++ pypinyin/contrib/tone_convert.py
@@ -1,7 +1,6 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-# 向后兼容
 from pypinyin.style._tone_convert import (  # noqa
     to_normal,
     to_tone,
@@ -9,6 +8,9 @@ from pypinyin.style._tone_convert import (  # noqa
     to_tone3,
     to_initials,
     to_finals,
+    to_finals_tone,
+    to_finals_tone2,
+    to_finals_tone3,
     tone_to_normal,
     tone_to_tone2,
     tone_to_tone3,
@@ -18,6 +20,7 @@ from pypinyin.style._tone_convert import (  # noqa
     tone3_to_normal,
     tone3_to_tone,
     tone3_to_tone2,
+    # 向后兼容
     _improve_tone3,
     _get_number_from_pinyin,
     _v_to_u,